fix: move options to a different file
